refactor : 특정 로직에만 사용하는 커스텀훅 같은 디렉토리 안으로 옮김
