+++ src/components/home/create/steps/Step3.tsx
@@ -3,7 +3,7 @@ import useBottomSheet from '@lib/hooks/useBottomSheet';
 import SelectMoney from '@components/common/bottomSheets/contractSheet/SelectMoney';
 import styled from 'styled-components';
 import useValidation, { TValidationResult } from '@lib/hooks/useValidation';
-import useStackAmount from '@lib/hooks/useStackAmount';
+import useStackAmount from '@components/home/create/utils/useStackAmount';
 import { useAppDispatch, useAppSelector } from '@store/app/hooks';
 import {
   selectStep3InitData,
@@ -102,7 +102,7 @@ function Step3({ onNextButtonClick }: CreateStepProps) {
             value={
               form.contractAmount === 0
                 ? ''
-                : getCommaThreeDigits(form.contractAmount)
+                : form.contractAmount.toLocaleString('ko-KR')
             }
             error={validateAmount.error}
             readonly={true}
