+++ src/components/mypage/KidsRecordItem.tsx
@@ -6,7 +6,7 @@ import OverViewData from './OverViewContent';
 
 function KidsRecordItem({ kid }: { kid: IKidListDTO }) {
   const overViewData = [
-    { name: '총 저금액', value: `${getCommaThreeDigits(kid.savings)}` },
+    { name: '총 저금액', value: `${kid.savings.toLocaleString('ko-KR')}` },
     {
       name: '아이의 총 돈길',
       value: kid.totalChallenge,
