+++ src/index.tsx
@@ -11,7 +11,6 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import './assets/fonts/fontStyle.css';
 import 'react-spring-bottom-sheet/dist/style.css';
-import './assets/fonts/fontStyle.css';
 import ReactModal from 'react-modal';
 import { ModalsContextProvider } from './components/common/modals/ModalsContext';
 import GlobalBottomSheet from '@components/common/bottomSheets/GlobalBottomSheet';
