--- src/lib/utils/get/getTypedStep.ts
@@ -1,9 +1,0 @@
-const getTypedStep = (parsedStep: number) => {
-  if (parsedStep > 0 && parsedStep <= 5) {
-    return parsedStep as 1 | 2 | 3 | 4 | 5;
-  } else {
-    throw 'step error';
-  }
-};
-
-export default getTypedStep;
